Overwrite the old version file.
