+++ lib/pages/setting/block/block_rule_edit_page.dart
@@ -1,15 +1,14 @@
 import 'package:fehviewer/common/controller/block_controller.dart';
 import 'package:fehviewer/common/service/layout_service.dart';
-import 'package:fehviewer/common/service/theme_service.dart';
 import 'package:fehviewer/component/setting_base.dart';
 import 'package:fehviewer/fehviewer.dart';
+import 'package:fehviewer/widget/cupertino/sliver_list_section.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:get/get.dart';
+import 'package:sliver_tools/sliver_tools.dart';
 
 class BlockRuleEditPage extends GetView<BlockController> {
-  const BlockRuleEditPage({super.key, this.blockRule});
-
-  final BlockRule? blockRule;
+  const BlockRuleEditPage({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -18,8 +17,7 @@ class BlockRuleEditPage extends GetView<BlockController> {
     final BlockRule? _blockRuleFromArg =
         Get.arguments is BlockRule ? Get.arguments as BlockRule : null;
 
-    BlockRule _blockRule = blockRule ??
-        _blockRuleFromArg ??
+    BlockRule _blockRule = _blockRuleFromArg ??
         BlockRule(
           ruleText: '',
           blockType: controller.latestBlockType?.name ?? BlockType.title.name,
@@ -30,129 +28,105 @@ class BlockRuleEditPage extends GetView<BlockController> {
     controller.blockRuleTextEditingController.text = _blockRule.ruleText ?? '';
     controller.currentEnableRegex = _blockRule.enableRegex ?? false;
 
-    final List<Widget> _list = <Widget>[
-      TextSwitchItem(
-        L10n.of(context).enable,
-        value: _blockRule.enabled ?? true,
-        onChanged: (val) {
-          _blockRule = _blockRule.copyWith(enabled: val);
-        },
+    return CupertinoPageScaffold(
+      backgroundColor: CupertinoColors.systemGroupedBackground,
+      navigationBar: CupertinoNavigationBar(
+        middle: Text(_title),
+        trailing: CupertinoButton(
+          padding: const EdgeInsets.all(0),
+          minSize: 40,
+          child: const Icon(
+            CupertinoIcons.check_mark_circled,
+            size: 28,
+          ),
+          onPressed: controller.isRegexFormatError
+              ? null
+              : () async {
+                  FocusScope.of(context).requestFocus(FocusNode());
+                  logger.d('_blockRule ${_blockRule.toJson()}');
+                  Get.back<BlockRule>(
+                    id: isLayoutLarge ? 2 : null,
+                    result: _blockRule,
+                  );
+                },
+        ),
       ),
-      TextSwitchItem(
-        L10n.of(context).regex,
-        value: _blockRule.enableRegex ?? false,
-        onChanged: (val) {
-          _blockRule = _blockRule.copyWith(enableRegex: val);
-          controller.currentEnableRegex = val;
-        },
-        // hideDivider: true,
-      ),
-      _BlockTypeSelector(
-        initValue: BlockType.values
-            .byName(_blockRule.blockType ?? BlockType.title.name),
-        onChanged: (BlockType value) {
-          controller.latestBlockType = value;
-          _blockRule = _blockRule.copyWith(blockType: value.name);
-        },
-      ),
-      GroupItem(
-        title: L10n.of(context).block_rule,
-        child: Container(
-          color: CupertinoDynamicColor.resolve(
-              ehTheme.itemBackgroundColor!, Get.context!),
-          constraints: const BoxConstraints(minHeight: kItemHeight),
-          padding: const EdgeInsets.symmetric(horizontal: 20),
-          child: SafeArea(
-            child: Builder(builder: (context) {
-              return Row(
+      child: CustomScrollView(slivers: [
+        SliverSafeArea(
+          sliver: MultiSliver(
+            children: [
+              SliverCupertinoListSection.listInsetGrouped(children: [
+                // enable switch
+                CupertinoListTile(
+                  title: Text(L10n.of(context).enable),
+                  trailing: StatefulBuilder(builder: (context, setState) {
+                    return CupertinoSwitch(
+                      value: _blockRule.enabled ?? true,
+                      onChanged: (val) {
+                        _blockRule = _blockRule.copyWith(enabled: val);
+                        setState(() {});
+                      },
+                    );
+                  }),
+                ),
+
+                // regex switch
+                CupertinoListTile(
+                  title: Text(L10n.of(context).regex),
+                  trailing: StatefulBuilder(builder: (context, setState) {
+                    return CupertinoSwitch(
+                      value: _blockRule.enableRegex ?? false,
+                      onChanged: (val) {
+                        _blockRule = _blockRule.copyWith(enableRegex: val);
+                        controller.currentEnableRegex = val;
+                        setState(() {});
+                      },
+                    );
+                  }),
+                ),
+                _BlockTypeSelector(
+                  initValue: BlockType.values
+                      .byName(_blockRule.blockType ?? BlockType.title.name),
+                  onChanged: (BlockType value) {
+                    controller.latestBlockType = value;
+                    _blockRule = _blockRule.copyWith(blockType: value.name);
+                  },
+                ),
+              ]),
+              SliverCupertinoListSection.listInsetGrouped(
+                header: Text(L10n.of(context).block_rule),
                 children: [
-                  Expanded(
-                    child: CupertinoTextField(
-                      decoration: null,
-                      maxLines: null,
-                      controller: controller.blockRuleTextEditingController,
-                      placeholder: L10n.of(context).block_rule,
-                      placeholderStyle: const TextStyle(
-                        fontWeight: FontWeight.w400,
-                        color: CupertinoColors.placeholderText,
-                        height: 1.25,
-                      ),
-                      textInputAction: TextInputAction.done,
-                      style: const TextStyle(height: 1.2),
-                      onEditingComplete: () {
-                        FocusScope.of(context).requestFocus(FocusNode());
-                      },
-                      keyboardType: TextInputType.text,
-                      onChanged: (value) {
-                        logger.t('value $value');
-                        _blockRule = _blockRule.copyWith(
-                            ruleText: value.replaceAll('\n', ' '));
-                      },
+                  CupertinoTextField(
+                    padding: const EdgeInsets.symmetric(
+                        horizontal: 20, vertical: 10),
+                    decoration: null,
+                    maxLines: null,
+                    controller: controller.blockRuleTextEditingController,
+                    placeholder: L10n.of(context).block_rule,
+                    placeholderStyle: const TextStyle(
+                      fontWeight: FontWeight.w400,
+                      color: CupertinoColors.placeholderText,
+                      height: 1.25,
                     ),
+                    textInputAction: TextInputAction.done,
+                    style: const TextStyle(height: 1.3),
+                    onEditingComplete: () {
+                      FocusScope.of(context).requestFocus(FocusNode());
+                    },
+                    keyboardType: TextInputType.text,
+                    onChanged: (value) {
+                      logger.t('value $value');
+                      _blockRule = _blockRule.copyWith(
+                          ruleText: value.replaceAll('\n', ' '));
+                    },
                   ),
                 ],
-              );
-            }),
+              ),
+            ],
           ),
         ),
-      ),
-      Obx(() {
-        return Container(
-          padding: const EdgeInsets.only(
-            left: 20,
-            bottom: 4,
-            top: 4,
-          ),
-          width: double.infinity,
-          child: Text(
-            controller.isRegexFormatError
-                ? L10n.of(context).regex_format_error
-                : '',
-            style: TextStyle(
-                fontSize: 14,
-                color: CupertinoDynamicColor.resolve(
-                    CupertinoColors.systemRed, context)),
-            textAlign: TextAlign.start,
-          ),
-        );
-      }),
-    ];
-
-    return Obx(() {
-      return CupertinoPageScaffold(
-        backgroundColor: !ehTheme.isDarkMode
-            ? CupertinoColors.secondarySystemBackground
-            : null,
-        navigationBar: CupertinoNavigationBar(
-          padding: const EdgeInsetsDirectional.only(end: 8),
-          middle: Text(_title),
-          trailing: CupertinoButton(
-            padding: const EdgeInsets.all(0),
-            minSize: 40,
-            child: const Icon(
-              CupertinoIcons.check_mark_circled,
-              size: 28,
-            ),
-            onPressed: controller.isRegexFormatError
-                ? null
-                : () async {
-                    FocusScope.of(context).requestFocus(FocusNode());
-                    logger.d('_blockRule ${_blockRule.toJson()}');
-                    Get.back<BlockRule>(
-                      id: isLayoutLarge ? 2 : null,
-                      result: _blockRule,
-                    );
-                  },
-          ),
-        ),
-        child: ListView.builder(
-          itemBuilder: (context, index) {
-            return _list[index];
-          },
-          itemCount: _list.length,
-        ),
-      );
-    });
+      ]),
+    );
   }
 }
 
@@ -201,13 +175,9 @@ class _BlockTypeSelectorState extends State<_BlockTypeSelector> {
   Widget build(BuildContext context) {
     return Container(
       width: double.infinity,
-      color: CupertinoDynamicColor.resolve(
-          ehTheme.itemBackgroundColor!, Get.context!),
       child: Container(
         padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 10),
-        constraints: const BoxConstraints(
-          minHeight: kItemHeight,
-        ),
+        constraints: const BoxConstraints(minHeight: kItemHeight),
         child: SafeArea(
           child: CupertinoSlidingSegmentedControl<BlockType>(
             children: <BlockType, Widget>{
