+++ lib/pages/setting/download_setting_page.dart
@@ -1,35 +1,34 @@
 import 'package:fehviewer/common/controller/download_controller.dart';
 import 'package:fehviewer/common/service/ehsetting_service.dart';
-import 'package:fehviewer/common/service/theme_service.dart';
-import 'package:fehviewer/component/setting_base.dart';
+import 'package:fehviewer/fehviewer.dart';
+import 'package:fehviewer/pages/setting/setting_items/selector_Item.dart';
+import 'package:fehviewer/widget/cupertino/sliver_list_section.dart';
 import 'package:file_picker/file_picker.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:get/get.dart';
 import 'package:shared_storage/shared_storage.dart' as ss;
-
-import '../../fehviewer.dart';
-import 'setting_items/selector_Item.dart';
+import 'package:sliver_tools/sliver_tools.dart';
 
 class DownloadSettingPage extends StatelessWidget {
+  const DownloadSettingPage({super.key});
+
   @override
   Widget build(BuildContext context) {
-    final Widget cps = Obx(() {
-      return CupertinoPageScaffold(
-          backgroundColor: !ehTheme.isDarkMode
-              ? CupertinoColors.secondarySystemBackground
-              : null,
-          navigationBar: CupertinoNavigationBar(
-            transitionBetweenRoutes: true,
-            middle: Text(L10n.of(context).download),
-          ),
-          child: ListViewDownloadSetting());
-    });
-
-    return cps;
+    return CupertinoPageScaffold(
+      backgroundColor: CupertinoColors.systemGroupedBackground,
+      navigationBar: CupertinoNavigationBar(
+        middle: Text(L10n.of(context).download),
+      ),
+      child: CustomScrollView(slivers: [
+        SliverSafeArea(sliver: DownloadSetting()),
+      ]),
+    );
   }
 }
 
-class ListViewDownloadSetting extends StatelessWidget {
+class DownloadSetting extends StatelessWidget {
+  DownloadSetting({super.key});
+
   final EhSettingService ehSettingService = Get.find();
   final DownloadController downloadController = Get.find();
 
@@ -40,88 +39,100 @@ class ListViewDownloadSetting extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final List<Widget> _list = <Widget>[
-      // 下载路径
-      if (!GetPlatform.isIOS)
-        Obx(() {
-          ehSettingService.downloadLocatino;
-          return FutureBuilder<String>(
-              future: defDownloadPath,
-              builder: (context, snapshot) {
-                late String path;
-                if (snapshot.connectionState == ConnectionState.done) {
-                  if (ehSettingService.downloadLocatino.isEmpty) {
-                    path = snapshot.data ?? '';
-                  } else {
-                    path = ehSettingService.downloadLocatino;
-                  }
-                } else {
-                  path = '';
-                }
-
-                return SelectorSettingItem(
-                  title: L10n.of(context).download_location,
-                  desc: path,
-                  suffix: CupertinoButton(
-                    padding: const EdgeInsets.all(0),
-                    minSize: 36,
-                    child: const Icon(CupertinoIcons.refresh),
-                    onPressed: () async => ehSettingService.downloadLocatino =
-                        await defDownloadPath,
-                  ),
-                  onTap: () async {
-                    if (GetPlatform.isAndroid) {
-                      // android 使用 SAF
-                      final uri = await ss.openDocumentTree();
-                      logger.d('uri $uri');
-                      if (uri != null) {
-                        ehSettingService.downloadLocatino = uri.toString();
-                      }
+    return MultiSliver(children: [
+      SliverCupertinoListSection.listInsetGrouped(children: [
+        // 下载路径
+        if (!GetPlatform.isIOS)
+          Obx(() {
+            ehSettingService.downloadLocatino;
+            return FutureBuilder<String>(
+                future: defDownloadPath,
+                builder: (context, snapshot) {
+                  late String path;
+                  if (snapshot.connectionState == ConnectionState.done) {
+                    if (ehSettingService.downloadLocatino.isEmpty) {
+                      path = snapshot.data ?? '';
                     } else {
-                      final String? result =
-                          await FilePicker.platform.getDirectoryPath();
-                      logger.d('set $result');
-
-                      if (result != null) {
-                        ehSettingService.downloadLocatino = result;
-                      }
+                      path = ehSettingService.downloadLocatino;
                     }
-                  },
-                );
-              });
-        }),
-      if (GetPlatform.isAndroid || GetPlatform.isFuchsia)
-        TextSwitchItem(
-          L10n.of(context).allow_media_scan,
-          value: ehSettingService.allowMediaScan,
-          onChanged: _handleAllowMediaScanChanged,
+                  } else {
+                    path = '';
+                  }
+
+                  return CupertinoListTile(
+                    padding: const EdgeInsetsDirectional.only(
+                        start: 20.0, end: 14.0, top: 6.0, bottom: 6.0),
+                    title: Text(L10n.of(context).download_location),
+                    subtitle: Text(path, maxLines: 10),
+                    trailing: CupertinoButton(
+                      padding: const EdgeInsets.all(0),
+                      minSize: 14,
+                      child: const Icon(CupertinoIcons.refresh),
+                      onPressed: () async => ehSettingService.downloadLocatino =
+                          await defDownloadPath,
+                    ),
+                    onTap: () async {
+                      if (GetPlatform.isAndroid) {
+                        // android 使用 SAF
+                        final uri = await ss.openDocumentTree();
+                        logger.d('uri $uri');
+                        if (uri != null) {
+                          ehSettingService.downloadLocatino = uri.toString();
+                        }
+                      } else {
+                        final String? result =
+                            await FilePicker.platform.getDirectoryPath();
+                        logger.d('set $result');
+
+                        if (result != null) {
+                          ehSettingService.downloadLocatino = result;
+                        }
+                      }
+                    },
+                  );
+                });
+          }),
+
+        // allow_media_scan
+        if (GetPlatform.isAndroid || GetPlatform.isFuchsia)
+          CupertinoListTile(
+            title: Text(L10n.of(context).allow_media_scan),
+            trailing: Obx(() {
+              return CupertinoSwitch(
+                value: ehSettingService.allowMediaScan,
+                onChanged: _handleAllowMediaScanChanged,
+              );
+            }),
+          ),
+
+        _buildPreloadImageItem(context),
+        _buildMultiDownloadItem(context),
+        _buildDownloadOrigImageItem(context),
+
+        //
+      ]),
+
+      // 数据处理
+      SliverCupertinoListSection.listInsetGrouped(children: [
+        // 恢复下载任务数据
+        CupertinoListTile(
+          title: Text(L10n.of(context).restore_tasks_data),
+          trailing: const CupertinoListTileChevron(),
+          onTap: () async {
+            final downloadController = Get.find<DownloadController>();
+            await downloadController.restoreGalleryTasks(init: true);
+          },
         ),
-      _buildPreloadImageItem(context),
-      _buildMultiDownloadItem(context),
-      _buildDownloadOrigImageItem(context),
-      // 恢复下载任务数据
-      SelectorSettingItem(
-        title: L10n.of(context).restore_tasks_data,
-        onTap: () async {
-          final downloadController = Get.find<DownloadController>();
-          await downloadController.restoreGalleryTasks(init: true);
-        },
-      ),
-      // 重建下载任务数据
-      SelectorSettingItem(
-        title: L10n.of(context).rebuild_tasks_data,
-        onTap: () async {
-          await Get.find<DownloadController>().rebuildGalleryTasks();
-        },
-        hideDivider: true,
-      ),
-    ];
-    return ListView.builder(
-      itemCount: _list.length,
-      itemBuilder: (BuildContext context, int index) {
-        return _list[index];
-      },
-    );
+        // 重建下载任务数据
+        CupertinoListTile(
+          title: Text(L10n.of(context).rebuild_tasks_data),
+          trailing: const CupertinoListTileChevron(),
+          onTap: () async {
+            await Get.find<DownloadController>().rebuildGalleryTasks();
+          },
+        ),
+      ]),
+    ]);
   }
 }
 
@@ -138,9 +149,8 @@ Widget _buildDownloadOrigImageItem(BuildContext context,
     DownloadOrigImageType.always: L10n.of(context).always,
   };
   return Obx(() {
-    return SelectorItem<DownloadOrigImageType>(
+    return SelectorCupertinoListTile<DownloadOrigImageType>(
       title: _title,
-      hideDivider: hideDivider,
       actionMap: modeMap,
       initVal: ehSettingService.downloadOrigType,
       onValueChanged: (val) => ehSettingService.downloadOrigType = val,
@@ -149,8 +159,7 @@ Widget _buildDownloadOrigImageItem(BuildContext context,
 }
 
 /// 预载图片数量
-Widget _buildPreloadImageItem(BuildContext context,
-    {bool hideDivider = false}) {
+Widget _buildPreloadImageItem(BuildContext context) {
   final String _title = L10n.of(context).preload_image;
   final EhSettingService ehSettingService = Get.find();
 
@@ -182,9 +191,10 @@ Widget _buildPreloadImageItem(BuildContext context,
         });
   }
 
-  return Obx(() => SelectorSettingItem(
-        title: _title,
-        selector: ehSettingService.preloadImage.toString(),
+  return Obx(() => CupertinoListTile(
+        title: Text(_title),
+        trailing: const CupertinoListTileChevron(),
+        additionalInfo: Text(ehSettingService.preloadImage.toString()),
         onTap: () async {
           final int? _result = await _showActionSheet(context);
           if (_result != null) {
@@ -195,7 +205,7 @@ Widget _buildPreloadImageItem(BuildContext context,
 }
 
 /// 同时下载图片数量
-Widget _buildMultiDownloadItem(BuildContext context, {bool hideLine = false}) {
+Widget _buildMultiDownloadItem(BuildContext context) {
   final String _title = L10n.of(context).multi_download;
   final EhSettingService ehSettingService = Get.find();
 
@@ -227,10 +237,10 @@ Widget _buildMultiDownloadItem(BuildContext context, {bool hideLine = false}) {
         });
   }
 
-  return Obx(() => SelectorSettingItem(
-        title: _title,
-        hideDivider: hideLine,
-        selector: ehSettingService.multiDownload.toString(),
+  return Obx(() => CupertinoListTile(
+        title: Text(_title),
+        trailing: const CupertinoListTileChevron(),
+        additionalInfo: Text(ehSettingService.multiDownload.toString()),
         onTap: () async {
           final int? _result = await _showActionSheet(context);
           if (_result != null) {
