+++ lib/pages/setting/eh_setting_page.dart
@@ -3,39 +3,36 @@ import 'package:fehviewer/common/controller/user_controller.dart';
 import 'package:fehviewer/common/service/ehsetting_service.dart';
 import 'package:fehviewer/common/service/layout_service.dart';
 import 'package:fehviewer/common/service/locale_service.dart';
-import 'package:fehviewer/common/service/theme_service.dart';
+import 'package:fehviewer/component/setting_base.dart';
 import 'package:fehviewer/fehviewer.dart';
 import 'package:fehviewer/network/api.dart';
 import 'package:fehviewer/network/request.dart';
 import 'package:fehviewer/pages/login/controller/login_controller.dart';
+import 'package:fehviewer/widget/cupertino/sliver_list_section.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:get/get.dart';
 import 'package:open_by_default/open_by_default.dart';
-
-import '../../component/setting_base.dart';
+import 'package:sliver_tools/sliver_tools.dart';
 
 class EhSettingPage extends StatelessWidget {
-  const EhSettingPage({Key? key}) : super(key: key);
+  const EhSettingPage({super.key});
 
   @override
   Widget build(BuildContext context) {
-    final Widget cps = Obx(() {
-      return CupertinoPageScaffold(
-          backgroundColor: !ehTheme.isDarkMode
-              ? CupertinoColors.secondarySystemBackground
-              : null,
-          navigationBar: CupertinoNavigationBar(
-            middle: Text(L10n.of(context).eh),
-          ),
-          child: ListViewEhSetting());
-    });
-
-    return cps;
+    return CupertinoPageScaffold(
+      backgroundColor: CupertinoColors.systemGroupedBackground,
+      navigationBar: CupertinoNavigationBar(
+        middle: Text(L10n.of(context).eh),
+      ),
+      child: CustomScrollView(slivers: [
+        SliverSafeArea(sliver: ListViewEhSetting()),
+      ]),
+    );
   }
 }
 
 class ListViewEhSetting extends StatelessWidget {
-  ListViewEhSetting({Key? key}) : super(key: key);
+  ListViewEhSetting({super.key});
 
   final EhSettingService _ehSettingService = Get.find();
   final UserController userController = Get.find();
@@ -43,203 +40,209 @@ class ListViewEhSetting extends StatelessWidget {
   final LocaleService localeService = Get.find();
   final LoginController loginController = Get.put(LoginController());
 
+  Future<void> _handleSiteChanged(bool newValue) async {
+    _ehSettingService.isSiteEx(newValue);
+    Global.forceRefreshUconfig = true;
+    if (newValue) {
+      getExIgneous();
+    }
+    Api.selEhProfile();
+    loginController.asyncGetUserInfo();
+  }
+
   @override
   Widget build(BuildContext context) {
-    final bool _favLongTap = _ehSettingService.isFavLongTap.value;
     final bool _isLogin = userController.isLogin;
-    final bool _isClipboar = _ehSettingService.isClipboardLink.value;
-
-    final bool _autoSelectProfile = _ehSettingService.autoSelectProfile;
-
-    Future<void> _handleSiteChanged(bool newValue) async {
-      _ehSettingService.isSiteEx(newValue);
-      Global.forceRefreshUconfig = true;
-      if (newValue) {
-        getExIgneous();
-      }
-      Api.selEhProfile();
-      loginController.asyncGetUserInfo();
-    }
-
-    void _handleFavLongTapChanged(bool newValue) {
-      _ehSettingService.isFavLongTap.value = newValue;
-    }
-
-    void _handleClipboarLinkTapChange(bool val) {
-      _ehSettingService.isClipboardLink.value = val;
-    }
-
     Future<EhHome?> _futureImageLimits = getEhHome(refresh: true);
 
-    final List<Widget> _list = <Widget>[
-      // if (_isLogin)
-      //   Obx(() {
-      //     return GestureDetector(
-      //       onLongPress: Api.selEhProfile,
-      //       child: TextSwitchItem(
-      //         L10n.of(context).galery_site,
-      //         intValue: _ehSettingService.isSiteEx.value,
-      //         onChanged: _handleSiteChanged,
-      //         desc: L10n.of(context).current_site('E-Hentai'),
-      //         descOn: L10n.of(context).current_site('ExHentai'),
-      //       ),
-      //     );
-      //   }),
-      if (_isLogin)
-        Obx(() {
-          return SlidingSegmentedItem<String>(
-            L10n.of(context).galery_site,
-            intValue: _ehSettingService.isSiteEx.value
-                ? EHConst.EX_BASE_HOST
-                : EHConst.EH_BASE_HOST,
-            onValueChanged: (val) {
-              logger.d('val  $val');
-              _handleSiteChanged(EHConst.EX_BASE_HOST == val);
-            },
-            slidingChildren: const {
-              EHConst.EH_BASE_HOST: Text('E-Hentai', textScaleFactor: 0.8),
-              EHConst.EX_BASE_HOST: Text('ExHentai', textScaleFactor: 0.8)
-            },
-          );
-        }),
-      TextSwitchItem(
-        L10n.of(context).link_redirect,
-        value: _ehSettingService.linkRedirect,
-        onChanged: (val) => _ehSettingService.linkRedirect = val,
-        desc: L10n.of(context).link_redirect_summary,
-      ),
-      TextSwitchItem(
-        L10n.of(context).redirect_thumb_link,
-        value: _ehSettingService.redirectThumbLink,
-        onChanged: (val) => _ehSettingService.redirectThumbLink = val,
-        desc: L10n.of(context).redirect_thumb_link_summary,
-      ),
-      if (_isLogin)
-        const SelectorSettingItem(
-          title: 'Cookie',
-          selector: '',
-          onTap: showUserCookie,
-        ),
-      TextSwitchItem(
-        L10n.of(context).auto_select_profile,
-        value: _autoSelectProfile,
-        hideDivider: !_isLogin,
-        onChanged: (val) => _ehSettingService.autoSelectProfile = val,
-      ),
-      if (_isLogin)
-        SelectorSettingItem(
-          title: L10n.of(context).ehentai_settings,
-          selector: L10n.of(context).setting_on_website,
-          onTap: () {
-            Get.toNamed(
-              EHRoutes.mySettings,
-              id: isLayoutLarge ? 2 : null,
-            );
-          },
-          onLongPress: () async {
-            await Api.selEhProfile();
-            showToast('set EhProfile succs');
-          },
-        ),
-      if (_isLogin)
-        SelectorSettingItem(
-          title: L10n.of(context).ehentai_my_tags,
-          selector: L10n.of(context).mytags_on_website,
-          onTap: () {
-            Get.toNamed(
-              EHRoutes.myTags,
-              id: isLayoutLarge ? 2 : null,
-            );
-          },
-        ),
-      if (_isLogin)
-        StatefulBuilder(builder: (context, setState) {
-          return FutureBuilder<EhHome?>(
-              future: _futureImageLimits,
-              initialData: hiveHelper.getEhHome(),
-              builder: (context, snapshot) {
-                EhHome? ehHome = snapshot.data;
-                if (snapshot.connectionState == ConnectionState.done) {
-                  if (ehHome != null) {
-                    hiveHelper.setEhHome(ehHome);
+    return MultiSliver(children: [
+      SliverCupertinoListSection.listInsetGrouped(children: [
+        if (_isLogin)
+          CupertinoListTile(
+            title: Text(L10n.of(context).galery_site),
+            trailing: Obx(() {
+              return CupertinoSlidingSegmentedControl<String>(
+                groupValue: _ehSettingService.isSiteEx.value
+                    ? EHConst.EX_BASE_HOST
+                    : EHConst.EH_BASE_HOST,
+                children: const {
+                  EHConst.EH_BASE_HOST: Text('E-Hentai', textScaleFactor: 0.8),
+                  EHConst.EX_BASE_HOST: Text('ExHentai', textScaleFactor: 0.8)
+                },
+                onValueChanged: (String? val) {
+                  if (val != null) {
+                    _handleSiteChanged(val == EHConst.EX_BASE_HOST);
                   }
-                }
-                return SelectorSettingItem(
-                  hideDivider: true,
-                  title: L10n.of(context).image_limits,
-                  selector: ehHome == null
-                      ? ''
-                      : '${ehHome.currentLimit ?? ''} / ${ehHome.totLimit ?? ''}',
-                  desc:
-                      '${L10n.of(context).reset_cost}: ${ehHome?.resetCost ?? 0} GP',
-                  suffix: snapshot.connectionState != ConnectionState.done
-                      ? const CupertinoActivityIndicator()
-                      : const SizedBox(),
-                  onTap: () {
-                    setState(() {
-                      _futureImageLimits = getEhHome(refresh: true);
-                    });
-                  },
-                );
-              });
-        }),
-      const ItemSpace(),
-      SelectorSettingItem(
-        title: 'WebDAV',
-        onTap: () {
-          Get.toNamed(
-            EHRoutes.webDavSetting,
-            id: isLayoutLarge ? 2 : null,
-          );
-        },
-        hideDivider: true,
-      ),
-      if (GetPlatform.isAndroid)
-        FutureBuilder<bool>(future: () async {
-          final _androidInfo = await deviceInfo.androidInfo;
-          return _androidInfo.version.sdkInt >= 31;
-        }(), builder: (context, snapshot) {
-          if (snapshot.connectionState == ConnectionState.done &&
-              (snapshot.data ?? false)) {
-            return Column(
-              children: [
-                const ItemSpace(),
-                SelectorSettingItem(
-                  title: L10n.of(context).open_supported_links,
-                  desc: L10n.of(context).open_supported_links_summary,
-                  onTap: OpenByDefault.open,
-                  hideDivider: true,
-                ),
-              ],
+                },
+              );
+            }),
+          ),
+        CupertinoListTile(
+          title: Text(L10n.of(context).link_redirect),
+          subtitle: Text(L10n.of(context).link_redirect_summary),
+          trailing: Obx(() {
+            return CupertinoSwitch(
+              value: _ehSettingService.linkRedirect,
+              onChanged: (bool val) {
+                _ehSettingService.linkRedirect = val;
+              },
             );
-          } else {
-            return const SizedBox.shrink();
-          }
-        }),
+          }),
+        ),
+        CupertinoListTile(
+          title: Text(L10n.of(context).redirect_thumb_link),
+          subtitle: Text(L10n.of(context).redirect_thumb_link_summary),
+          trailing: Obx(() {
+            return CupertinoSwitch(
+              value: _ehSettingService.redirectThumbLink,
+              onChanged: (bool val) {
+                _ehSettingService.redirectThumbLink = val;
+              },
+            );
+          }),
+        ),
+        if (_isLogin)
+          const CupertinoListTile(
+            title: Text('Cookie'),
+            trailing: CupertinoListTileChevron(),
+            onTap: showUserCookie,
+          ),
+        CupertinoListTile(
+          title: Text(L10n.of(context).auto_select_profile),
+          trailing: Obx(() {
+            return CupertinoSwitch(
+              value: _ehSettingService.autoSelectProfile,
+              onChanged: (bool val) =>
+                  _ehSettingService.autoSelectProfile = val,
+            );
+          }),
+        ),
+        if (_isLogin)
+          CupertinoListTile(
+            title: Text(L10n.of(context).ehentai_settings),
+            subtitle: Text(L10n.of(context).setting_on_website),
+            trailing: const CupertinoListTileChevron(),
+            onTap: () {
+              Get.toNamed(
+                EHRoutes.mySettings,
+                id: isLayoutLarge ? 2 : null,
+              );
+            },
+          ),
+        if (_isLogin)
+          CupertinoListTile(
+            title: Text(L10n.of(context).ehentai_my_tags),
+            subtitle: Text(L10n.of(context).mytags_on_website),
+            trailing: const CupertinoListTileChevron(),
+            onTap: () {
+              Get.toNamed(
+                EHRoutes.myTags,
+                id: isLayoutLarge ? 2 : null,
+              );
+            },
+          ),
+        if (_isLogin)
+          StatefulBuilder(builder: (context, setState) {
+            return FutureBuilder<EhHome?>(
+                future: _futureImageLimits,
+                initialData: hiveHelper.getEhHome(),
+                builder: (context, snapshot) {
+                  EhHome? ehHome = snapshot.data;
+                  if (snapshot.connectionState == ConnectionState.done) {
+                    if (ehHome != null) {
+                      hiveHelper.setEhHome(ehHome);
+                    }
+                  }
+                  return CupertinoListTile(
+                    title: Text(L10n.of(context).image_limits),
+                    additionalInfo: Text(ehHome == null
+                        ? ''
+                        : '${ehHome.currentLimit ?? ''} / ${ehHome.totLimit ?? ''}'),
+                    subtitle: Text(
+                        '${L10n.of(context).reset_cost}: ${ehHome?.resetCost ?? 0} GP'),
+                    trailing: snapshot.connectionState != ConnectionState.done
+                        ? CupertinoActivityIndicator(
+                            radius: (CupertinoTheme.of(context)
+                                        .textTheme
+                                        .textStyle
+                                        .fontSize ??
+                                    14) /
+                                2,
+                          )
+                        : const CupertinoListTileChevron(),
+                    onTap: () {
+                      setState(() {
+                        _futureImageLimits = getEhHome(refresh: true);
+                      });
+                    },
+                  );
+                });
+          }),
+      ]),
 
-      const ItemSpace(),
-      TextSwitchItem(
-        L10n.of(context).default_favorites,
-        value: _favLongTap,
-        onChanged: _handleFavLongTapChanged,
-        desc: L10n.of(context).manually_sel_favorites,
-        descOn: L10n.of(context).last_favorites,
-      ),
+      // 云服务
+      SliverCupertinoListSection.listInsetGrouped(children: [
+        CupertinoListTile(
+          title: Text('WebDAV'),
+          trailing: const CupertinoListTileChevron(),
+          onTap: () {
+            Get.toNamed(
+              EHRoutes.webDavSetting,
+              id: isLayoutLarge ? 2 : null,
+            );
+          },
+        ),
+      ]),
 
-      TextSwitchItem(
-        L10n.of(context).clipboard_detection,
-        value: _isClipboar,
-        onChanged: _handleClipboarLinkTapChange,
-        desc: L10n.of(context).clipboard_detection_desc,
-        hideDivider: true,
-      ),
-    ];
+      // 默认打开
+      FutureBuilder<bool>(future: () async {
+        return GetPlatform.isAndroid &&
+            (await deviceInfo.androidInfo).version.sdkInt >= 31;
+      }(), builder: (context, snapshot) {
+        if (snapshot.connectionState == ConnectionState.done &&
+            (snapshot.data ?? false)) {
+          return SliverCupertinoListSection.listInsetGrouped(
+            children: [
+              CupertinoListTile(
+                title: Text(L10n.of(context).open_supported_links),
+                trailing: const CupertinoListTileChevron(),
+                // subtitle: Text(L10n.of(context).open_supported_links_summary),
+                onTap: () {
+                  OpenByDefault.open();
+                },
+              ),
+            ],
+            footer: Text(L10n.of(context).open_supported_links_summary),
+          );
+        } else {
+          return const SizedBox.shrink();
+        }
+      }),
 
-    return ListView.builder(
-      itemCount: _list.length,
-      itemBuilder: (BuildContext context, int index) {
-        return _list[index];
-      },
-    );
+      SliverCupertinoListSection.listInsetGrouped(children: [
+        CupertinoListTile(
+          title: Text(L10n.of(context).default_favorites),
+          subtitle: Text(L10n.of(context).manually_sel_favorites),
+          trailing: Obx(() {
+            return CupertinoSwitch(
+              value: _ehSettingService.isFavLongTap.value,
+              onChanged: (bool val) =>
+                  _ehSettingService.isFavLongTap.value = val,
+            );
+          }),
+        ),
+        CupertinoListTile(
+          title: Text(L10n.of(context).clipboard_detection),
+          subtitle: Text(L10n.of(context).clipboard_detection_desc),
+          trailing: Obx(() {
+            return CupertinoSwitch(
+              value: _ehSettingService.isClipboardLink.value,
+              onChanged: (bool val) =>
+                  _ehSettingService.isClipboardLink.value = val,
+            );
+          }),
+        ),
+      ]),
+    ]);
   }
 }
