+++ lib/pages/setting/layout_setting_page.dart
@@ -1,41 +1,38 @@
+import 'package:fehviewer/common/controller/tag_trans_controller.dart';
+import 'package:fehviewer/common/controller/user_controller.dart';
+import 'package:fehviewer/common/service/ehsetting_service.dart';
+import 'package:fehviewer/common/service/layout_service.dart';
+import 'package:fehviewer/common/service/locale_service.dart';
 import 'package:fehviewer/common/service/theme_service.dart';
+import 'package:fehviewer/const/theme_colors.dart';
+import 'package:fehviewer/fehviewer.dart';
+import 'package:fehviewer/pages/login/controller/login_controller.dart';
 import 'package:fehviewer/pages/setting/setting_items/selector_Item.dart';
+import 'package:fehviewer/widget/cupertino/sliver_list_section.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/foundation.dart';
 import 'package:get/get.dart';
-
-import '../../common/controller/tag_trans_controller.dart';
-import '../../common/controller/user_controller.dart';
-import '../../common/service/ehsetting_service.dart';
-import '../../common/service/layout_service.dart';
-import '../../common/service/locale_service.dart';
-import '../../component/setting_base.dart';
-import '../../const/theme_colors.dart';
-import '../../fehviewer.dart';
-import '../login/controller/login_controller.dart';
+import 'package:sliver_tools/sliver_tools.dart';
 
 class LayoutSettingPage extends StatelessWidget {
-  const LayoutSettingPage({Key? key}) : super(key: key);
+  const LayoutSettingPage({super.key});
 
   @override
   Widget build(BuildContext context) {
-    final Widget cps = Obx(() {
-      return CupertinoPageScaffold(
-          backgroundColor: !ehTheme.isDarkMode
-              ? CupertinoColors.secondarySystemBackground
-              : null,
-          navigationBar: CupertinoNavigationBar(
-            middle: Text(L10n.of(context).layout),
-          ),
-          child: ListViewLayoutSetting());
-    });
-
-    return cps;
+    return CupertinoPageScaffold(
+      backgroundColor: CupertinoColors.systemGroupedBackground,
+      navigationBar: CupertinoNavigationBar(
+        middle: Text(L10n.of(context).layout),
+      ),
+      child: CustomScrollView(slivers: [
+        SliverSafeArea(sliver: LayoutSettingList()),
+      ]),
+    );
   }
 }
 
-class ListViewLayoutSetting extends StatelessWidget {
-  ListViewLayoutSetting({Key? key}) : super(key: key);
+class LayoutSettingList extends StatelessWidget {
+  LayoutSettingList({super.key});
 
   final EhSettingService _ehSettingService = Get.find();
   final UserController userController = Get.find();
@@ -45,29 +42,32 @@ class ListViewLayoutSetting extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    void _handleGalleryListImgBlurChanged(bool newValue) {
-      _ehSettingService.isGalleryImgBlur.value = newValue;
-    }
+    return MultiSliver(children: [
+      SliverCupertinoListSection.listInsetGrouped(children: [
+        _buildThemeItem(context),
+        // dark_mode_effect
+        CupertinoListTile(
+          title: Text(L10n.of(context).dark_mode_effect),
+          trailing: Obx(() {
+            return CupertinoSlidingSegmentedControl<bool>(
+              groupValue: _ehSettingService.isPureDarkTheme.value,
+              children: {
+                false: Text(L10n.of(context).gray_black, textScaleFactor: 0.8),
+                true: Text(L10n.of(context).pure_black, textScaleFactor: 0.8)
+              },
+              onValueChanged: (bool? val) {
+                if (val != null) {
+                  _ehSettingService.isPureDarkTheme.value = val;
+                }
+              },
+            );
+          }),
+        ),
 
-    void _handlePureDarkChanged(bool newValue) {
-      _ehSettingService.isPureDarkTheme.value = newValue;
-    }
-
-    final List<Widget> _list = <Widget>[
-      _buildThemeItem(context),
-      Obx(() => TextSwitchItem(
-            L10n.of(context).dark_mode_effect,
-            value: _ehSettingService.isPureDarkTheme.value,
-            onChanged: _handlePureDarkChanged,
-            desc: L10n.of(context).gray_black,
-            descOn: L10n.of(context).pure_black,
-          )),
-      if (context.isTablet) _buildTableLayoutItem(context),
-      if (!Get.find<EhSettingService>().isSafeMode.value)
-        SelectorSettingItem(
-          hideDivider: true,
-          title: L10n.of(context).tabbar_setting,
-          selector: '',
+        // tabbar_setting
+        CupertinoListTile(
+          title: Text(L10n.of(context).tabbar_setting),
+          trailing: const CupertinoListTileChevron(),
           onTap: () {
             Get.toNamed(
               EHRoutes.pageSetting,
@@ -75,169 +75,223 @@ class ListViewLayoutSetting extends StatelessWidget {
             );
           },
         ),
-      const ItemSpace(),
-      if (localeService.isLanguageCodeZh)
-        Obx(() {
-          return SelectorSettingItem(
-            title: '标签翻译',
-            onTap: () {
-              Get.toNamed(
-                EHRoutes.tagTranslate,
-                id: isLayoutLarge ? 2 : null,
+      ]),
+      SliverCupertinoListSection.listInsetGrouped(
+        children: [
+          if (localeService.isLanguageCodeZh)
+            CupertinoListTile(
+              title: const Text('标签翻译'),
+              subtitle: Text('当前版本:${_ehSettingService.tagTranslatVer.value}'),
+              additionalInfo: Text(_ehSettingService.isTagTranslate
+                  ? L10n.of(context).on
+                  : L10n.of(context).off),
+              trailing: const CupertinoListTileChevron(),
+              onTap: () {
+                Get.toNamed(
+                  EHRoutes.tagTranslate,
+                  id: isLayoutLarge ? 2 : null,
+                );
+              },
+            ),
+          // japanese_title_in_gallery
+          CupertinoListTile(
+            title: Text(L10n.of(context).japanese_title_in_gallery),
+            subtitle: Text(L10n.of(context).japanese_title_in_gallery_summary),
+            trailing: Obx(() {
+              return CupertinoSwitch(
+                value: _ehSettingService.jpnTitleInGalleryPage,
+                onChanged: (bool val) {
+                  _ehSettingService.jpnTitleInGalleryPage = val;
+                },
               );
-            },
-            selector: _ehSettingService.isTagTranslate
-                ? L10n.of(context).on
-                : L10n.of(context).off,
-            desc: '当前版本:${_ehSettingService.tagTranslatVer.value}',
-          );
-        }),
-      TextSwitchItem(
-        L10n.of(context).japanese_title_in_gallery,
-        desc: L10n.of(context).japanese_title_in_gallery_summary,
-        value: _ehSettingService.jpnTitleInGalleryPage,
-        onChanged: (val) => _ehSettingService.jpnTitleInGalleryPage = val,
-      ),
-      // showComments
-      TextSwitchItem(
-        L10n.of(context).show_comments,
-        value: _ehSettingService.showComments,
-        onChanged: (val) => _ehSettingService.showComments = val,
-      ),
-      // showOnlyUploaderComment
-      Obx(() {
-        return AnimatedCrossFade(
-          firstChild: const SizedBox(),
-          secondChild: TextSwitchItem(
-            L10n.of(context).show_only_uploader_comment,
-            value: _ehSettingService.showOnlyUploaderComment,
-            onChanged: (val) => _ehSettingService.showOnlyUploaderComment = val,
+            }),
           ),
-          crossFadeState: _ehSettingService.showComments
-              ? CrossFadeState.showSecond
-              : CrossFadeState.showFirst,
-          secondCurve: Curves.easeOut,
-          duration: const Duration(milliseconds: 200),
-        );
-      }),
-      // showGalleryTags
-      TextSwitchItem(
-        L10n.of(context).show_gallery_tags,
-        value: _ehSettingService.showGalleryTags,
-        onChanged: (val) => _ehSettingService.showGalleryTags = val,
+
+          // showComments switch
+          CupertinoListTile(
+            title: Text(L10n.of(context).show_comments),
+            trailing: Obx(() {
+              return CupertinoSwitch(
+                value: _ehSettingService.showComments,
+                onChanged: (bool val) {
+                  _ehSettingService.showComments = val;
+                },
+              );
+            }),
+          ),
+
+          Obx(() {
+            return AnimatedCrossFade(
+              firstChild: const SizedBox(),
+              secondChild: CupertinoListTile(
+                title: Text(L10n.of(context).show_only_uploader_comment),
+                trailing: Obx(() {
+                  return CupertinoSwitch(
+                    value: _ehSettingService.showOnlyUploaderComment,
+                    onChanged: _ehSettingService.showComments
+                        ? (bool val) {
+                            _ehSettingService.showOnlyUploaderComment = val;
+                          }
+                        : null,
+                  );
+                }),
+              ),
+              crossFadeState: _ehSettingService.showComments
+                  ? CrossFadeState.showSecond
+                  : CrossFadeState.showFirst,
+              secondCurve: Curves.easeOut,
+              duration: const Duration(milliseconds: 200),
+            );
+          }),
+
+          // showGalleryTags switch
+          CupertinoListTile(
+            title: Text(L10n.of(context).show_gallery_tags),
+            trailing: Obx(() {
+              return CupertinoSwitch(
+                value: _ehSettingService.showGalleryTags,
+                onChanged: (bool val) {
+                  _ehSettingService.showGalleryTags = val;
+                },
+              );
+            }),
+          ),
+
+          // hideGalleryThumbnails switch
+          CupertinoListTile(
+            title: Text(L10n.of(context).hide_gallery_thumbnails),
+            trailing: Obx(() {
+              return CupertinoSwitch(
+                value: _ehSettingService.hideGalleryThumbnails,
+                onChanged: (bool val) {
+                  _ehSettingService.hideGalleryThumbnails = val;
+                },
+              );
+            }),
+          ),
+
+          // horizontalThumbnails
+          CupertinoListTile(
+            title: Text(L10n.of(context).horizontal_thumbnails),
+            trailing: Obx(() {
+              return CupertinoSwitch(
+                value: _ehSettingService.horizontalThumbnails,
+                onChanged: (bool val) {
+                  _ehSettingService.horizontalThumbnails = val;
+                },
+              );
+            }),
+          ),
+        ],
       ),
-      // hideGalleryThumbnails
-      TextSwitchItem(
-        L10n.of(context).hide_gallery_thumbnails,
-        value: _ehSettingService.hideGalleryThumbnails,
-        onChanged: (val) => _ehSettingService.hideGalleryThumbnails = val,
-      ),
-      // horizontalThumbnails
-      TextSwitchItem(
-        L10n.of(context).horizontal_thumbnails,
-        value: _ehSettingService.horizontalThumbnails,
-        onChanged: (val) => _ehSettingService.horizontalThumbnails = val,
-        hideDivider: true,
-      ),
-      const ItemSpace(),
-      TextSwitchItem(
-        L10n.of(context).hide_top_bar_on_scroll,
-        value: _ehSettingService.hideTopBarOnScroll,
-        onChanged: (val) => _ehSettingService.hideTopBarOnScroll = val,
-        // hideDivider: !localeService.isLanguageCodeZh,
-      ),
-      if (localeService.isLanguageCodeZh)
-        TextSwitchItem(
-          '画廊列表封面模糊处理',
-          value: _ehSettingService.isGalleryImgBlur.value,
-          onChanged: _handleGalleryListImgBlurChanged,
-          // hideDivider: true,
-          // desc: '画廊列表封面模糊效果',
+      SliverCupertinoListSection.listInsetGrouped(children: [
+        // hide_top_bar_on_scroll switch
+        CupertinoListTile(
+          title: Text(L10n.of(context).hide_top_bar_on_scroll),
+          trailing: Obx(() {
+            return CupertinoSwitch(
+              value: _ehSettingService.hideTopBarOnScroll,
+              onChanged: (bool val) {
+                _ehSettingService.hideTopBarOnScroll = val;
+              },
+            );
+          }),
         ),
-      _buildListModeItem(context),
-      SelectorSettingItem(
-        title: L10n.of(context).custom_width,
-        onTap: () {
-          NavigatorUtil.goItemWidthSettingPage();
-        },
-      ),
-      _buildTagLimitItem(context),
-      Obx(() {
-        return AnimatedCrossFade(
-          alignment: Alignment.center,
-          crossFadeState: _ehSettingService.listMode.value == ListModeEnum.list
-              ? CrossFadeState.showSecond
-              : CrossFadeState.showFirst,
-          firstCurve: Curves.easeIn,
-          secondCurve: Curves.easeOut,
-          duration: const Duration(milliseconds: 200),
-          firstChild: const SizedBox(),
-          secondChild: TextSwitchItem(
-            L10n.of(context).blurring_cover_background,
-            value: _ehSettingService.blurringOfCoverBackground,
-            onChanged: (val) =>
-                _ehSettingService.blurringOfCoverBackground = val,
-            hideDivider: _ehSettingService.listMode.value != ListModeEnum.list,
+
+        // isGalleryImgBlur
+        if (localeService.isLanguageCodeZh)
+          CupertinoListTile(
+            title: Text('画廊列表封面模糊处理'),
+            trailing: Obx(() {
+              return CupertinoSwitch(
+                value: _ehSettingService.isGalleryImgBlur.value,
+                onChanged: (bool val) {
+                  _ehSettingService.isGalleryImgBlur.value = val;
+                },
+              );
+            }),
           ),
-        );
-      }),
-      Obx(() {
-        return AnimatedCrossFade(
-          alignment: Alignment.center,
-          crossFadeState: _ehSettingService.listMode.value == ListModeEnum.list
-              ? CrossFadeState.showSecond
-              : CrossFadeState.showFirst,
-          firstCurve: Curves.easeIn,
-          secondCurve: Curves.easeOut,
-          duration: const Duration(milliseconds: 200),
-          firstChild: const SizedBox(),
-          secondChild: TextSwitchItem(
-            L10n.of(context).fixed_height_of_list_items,
-            value: _ehSettingService.fixedHeightOfListItems,
-            onChanged: (val) => _ehSettingService.fixedHeightOfListItems = val,
-            hideDivider: true,
-          ),
-        );
-      }),
-      const ItemSpace(),
-      Obx(() {
-        return SelectorSettingItem(
-          title: L10n.of(context).avatar,
-          hideDivider: !localeService.isLanguageCodeZh || !GetPlatform.isMobile,
+
+        _buildListModeItem(context),
+
+        // custom_width
+        CupertinoListTile(
+          title: Text(L10n.of(context).custom_width),
+          trailing: const CupertinoListTileChevron(),
+          onTap: () {
+            Get.toNamed(
+              EHRoutes.itemWidthSetting,
+              id: isLayoutLarge ? 2 : null,
+            );
+          },
+        ),
+
+        _buildTagLimitItem(context),
+
+        // blurring_cover_background switch
+        CupertinoListTile(
+          title: Text(L10n.of(context).blurring_cover_background),
+          trailing: Obx(() {
+            return CupertinoSwitch(
+              value: _ehSettingService.blurringOfCoverBackground,
+              onChanged: _ehSettingService.listMode.value == ListModeEnum.list
+                  ? (bool val) {
+                      _ehSettingService.blurringOfCoverBackground = val;
+                    }
+                  : null,
+            );
+          }),
+        ),
+
+        // fixed_height_of_list_items switch
+        CupertinoListTile(
+          title: Text(L10n.of(context).fixed_height_of_list_items),
+          trailing: Obx(() {
+            return CupertinoSwitch(
+              value: _ehSettingService.fixedHeightOfListItems,
+              onChanged: _ehSettingService.listMode.value == ListModeEnum.list
+                  ? (bool val) {
+                      _ehSettingService.fixedHeightOfListItems = val;
+                    }
+                  : null,
+            );
+          }),
+        ),
+      ]),
+      SliverCupertinoListSection.listInsetGrouped(children: [
+        // to avatar
+        CupertinoListTile(
+          title: Text(L10n.of(context).avatar),
+          trailing: const CupertinoListTileChevron(),
           onTap: () {
             Get.toNamed(
               EHRoutes.avatarSetting,
               id: isLayoutLarge ? 2 : null,
             );
           },
-          selector: _ehSettingService.showCommentAvatar
-              ? L10n.of(context).on
-              : L10n.of(context).off,
-        );
-      }),
-      if (localeService.isLanguageCodeZh && GetPlatform.isMobile)
-        TextSwitchItem(
-          '评论机翻按钮',
-          value: _ehSettingService.commentTrans.value,
-          onChanged: (bool newValue) =>
-              _ehSettingService.commentTrans.value = newValue,
-          desc: '关闭',
-          descOn: '用机器翻译将评论翻译为简体中文',
-          hideDivider: true,
         ),
-    ];
 
-    return ListView.builder(
-      itemCount: _list.length,
-      itemBuilder: (BuildContext context, int index) {
-        return _list[index];
-      },
-    );
+        // commentTrans switch
+        if (localeService.isLanguageCodeZh && GetPlatform.isMobile)
+          CupertinoListTile(
+            title: const Text('评论机翻按钮'),
+            subtitle: const Text('用机器翻译将评论翻译为简体中文'),
+            trailing: Obx(() {
+              return CupertinoSwitch(
+                value: _ehSettingService.commentTrans.value,
+                onChanged: (bool val) {
+                  _ehSettingService.commentTrans.value = val;
+                },
+              );
+            }),
+          ),
+      ]),
+    ]);
   }
 }
 
 /// 主题设置部件
-Widget _buildThemeItem(BuildContext context, {bool hideLine = false}) {
+Widget _buildThemeItem(BuildContext context) {
   final String _title = L10n.of(context).theme;
   final ThemeService themeService = Get.find();
 
@@ -248,9 +302,8 @@ Widget _buildThemeItem(BuildContext context, {bool hideLine = false}) {
   };
 
   return Obx(() {
-    return SelectorItem<ThemesModeEnum>(
+    return SelectorCupertinoListTile<ThemesModeEnum>(
       title: _title,
-      hideDivider: hideLine,
       actionMap: themeMap,
       initVal: themeService.themeModel,
       onValueChanged: (val) => themeService.themeModel = val,
@@ -259,7 +312,7 @@ Widget _buildThemeItem(BuildContext context, {bool hideLine = false}) {
 }
 
 /// 列表模式切换
-Widget _buildListModeItem(BuildContext context, {bool hideDivider = false}) {
+Widget _buildListModeItem(BuildContext context) {
   final String _title = L10n.of(context).list_mode;
   final EhSettingService ehSettingService = Get.find();
 
@@ -273,9 +326,8 @@ Widget _buildListModeItem(BuildContext context, {bool hideDivider = false}) {
       ListModeEnum.debugSimple: 'debugSimple',
   };
   return Obx(() {
-    return SelectorItem<ListModeEnum>(
+    return SelectorCupertinoListTile<ListModeEnum>(
       title: _title,
-      hideDivider: hideDivider,
       actionMap: modeMap,
       initVal: ehSettingService.listMode.value,
       onValueChanged: (val) => ehSettingService.listMode.value = val,
@@ -284,7 +336,7 @@ Widget _buildListModeItem(BuildContext context, {bool hideDivider = false}) {
 }
 
 /// 平板布局
-Widget _buildTableLayoutItem(BuildContext context, {bool hideLine = false}) {
+Widget _buildTableLayoutItem(BuildContext context) {
   final String _title = L10n.of(context).tablet_layout;
   final EhSettingService ehSettingService = Get.find();
 
@@ -295,9 +347,8 @@ Widget _buildTableLayoutItem(BuildContext context, {bool hideLine = false}) {
   };
 
   return Obx(() {
-    return SelectorItem<TabletLayout>(
+    return SelectorCupertinoListTile<TabletLayout>(
       title: _title,
-      hideDivider: hideLine,
       actionMap: localeMap,
       initVal: ehSettingService.tabletLayoutType,
       onValueChanged: (val) => ehSettingService.tabletLayoutType = val,
@@ -306,7 +357,7 @@ Widget _buildTableLayoutItem(BuildContext context, {bool hideLine = false}) {
 }
 
 /// tag上限
-Widget _buildTagLimitItem(BuildContext context, {bool hideDivider = false}) {
+Widget _buildTagLimitItem(BuildContext context) {
   final String _title = L10n.of(context).tag_limit;
   final EhSettingService ehSettingService = Get.find();
 
@@ -318,9 +369,8 @@ Widget _buildTagLimitItem(BuildContext context, {bool hideDivider = false}) {
   }
 
   return Obx(() {
-    return SelectorItem<int>(
+    return SelectorCupertinoListTile<int>(
       title: _title,
-      hideDivider: hideDivider,
       actionMap: modeMap,
       initVal: ehSettingService.listViewTagLimit,
       onValueChanged: (val) => ehSettingService.listViewTagLimit = val,
