+++ lib/pages/setting/tabbar_setting_page.dart
@@ -1,62 +1,38 @@
-import 'package:fehviewer/common/service/theme_service.dart';
-import 'package:fehviewer/component/setting_base.dart';
-import 'package:fehviewer/generated/l10n.dart';
+import 'package:fehviewer/fehviewer.dart';
 import 'package:fehviewer/pages/setting/controller/tab_setting_controller.dart';
 import 'package:fehviewer/pages/tab/controller/tabhome_controller.dart';
+import 'package:fehviewer/widget/cupertino/sliver_list_section.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:get/get.dart';
 import 'package:reorderables/reorderables.dart';
 import 'package:sliver_tools/sliver_tools.dart';
 
 class TabbarSettingPage extends StatelessWidget {
+  const TabbarSettingPage({super.key});
+
   @override
   Widget build(BuildContext context) {
-    return Obx(() {
-      return CupertinoPageScaffold(
-        backgroundColor: !ehTheme.isDarkMode
-            ? CupertinoColors.secondarySystemBackground
-            : null,
-        navigationBar: CupertinoNavigationBar(
-          middle: Text(L10n.of(context).tabbar_setting),
-        ),
-        child: CustomScrollView(
-          controller: Get.find<TabSettingController>().scrollController,
-          physics: const AlwaysScrollableScrollPhysics(),
-          slivers: [
-            SliverSafeArea(
-              left: false,
-              right: false,
-              sliver: MultiSliver(children: [
-                const TablistView(),
-                Container(
-                  padding: const EdgeInsets.only(
-                    left: 20,
-                    top: 4,
-                    bottom: 10,
-                    right: 20,
-                  ),
-                  width: double.infinity,
-                  child: Text(
-                    L10n.of(context).tab_sort,
-                    style: TextStyle(
-                      fontSize: kSummaryFontSize,
-                      color: CupertinoDynamicColor.resolve(
-                          CupertinoColors.secondaryLabel, context),
-                    ),
-                    textAlign: TextAlign.start,
-                  ),
-                ),
-              ]),
+    return CupertinoPageScaffold(
+      backgroundColor: CupertinoColors.systemGroupedBackground,
+      navigationBar: CupertinoNavigationBar(
+        middle: Text(L10n.of(context).tabbar_setting),
+      ),
+      child: CustomScrollView(slivers: [
+        SliverSafeArea(
+          sliver: MultiSliver(children: [
+            SliverCupertinoListSection.listInsetGrouped(
+              footer: Text(L10n.of(context).tab_sort),
+              children: const [TabbarListView()],
             ),
-          ],
+          ]),
         ),
-      );
-    });
+      ]),
+    );
   }
 }
 
-class TablistView extends StatelessWidget {
-  const TablistView({
+class TabbarListView extends StatelessWidget {
+  const TabbarListView({
     Key? key,
   }) : super(key: key);
 
@@ -64,33 +40,63 @@ class TablistView extends StatelessWidget {
   Widget build(BuildContext context) {
     final TabSettingController controller = Get.put(TabSettingController());
     return Obx(() {
-      return ReorderableSliverList(
-        delegate: ReorderableSliverChildListDelegate(controller.tabList
-            .map(
-              (e) => TextSwitchItem(
-                tabPages.tabTitles[e] ?? '',
-                key: UniqueKey(),
-                hideDivider: e == controller.tabList.last,
-                icon: Padding(
-                  padding: const EdgeInsets.only(right: 18.0),
-                  child: Icon(
-                    tabPages.iconDatas[e],
-                    color: CupertinoDynamicColor.resolve(
-                        CupertinoColors.systemGrey, Get.context!),
-                  ),
-                ),
-                iconIndent: 32,
-                value: controller.tabMap[e] ?? false,
-                onChanged: (controller.disableSwitch && controller.tabMap[e]!)
-                    ? null
-                    : (bool val) {
-                        controller.onChanged(val, e);
-                      },
-              ),
-            )
-            .toList()),
+      return ReorderableColumn(
         onReorder: controller.onReorder,
+        children: _buildList(context),
       );
     });
+  }
+
+  List<Widget> _buildList(BuildContext context) {
+    final TabSettingController controller = Get.put(TabSettingController());
+    final List<Widget> _list = [];
+
+    final Widget shortDivider = Container(
+      margin: const EdgeInsetsDirectional.only(start: 60),
+      color: CupertinoColors.separator.resolveFrom(context),
+      height: 0.5,
+    );
+
+    Widget _buildCupertinoListTile(String tag) {
+      return CupertinoListTile(
+        title: Text(tabPages.tabTitles[tag] ?? ''),
+        key: ValueKey(tag),
+        leading: Icon(
+          tabPages.iconDatas[tag],
+          color: CupertinoDynamicColor.resolve(
+              CupertinoColors.systemGrey, context),
+        ),
+        trailing: Obx(() {
+          return CupertinoSwitch(
+            value: controller.tabMap[tag] ?? false,
+            onChanged: (controller.disableSwitch && controller.tabMap[tag]!)
+                ? null
+                : (bool val) {
+                    controller.onChanged(val, tag);
+                  },
+          );
+        }),
+      );
+    }
+
+    // for in controller.tabList -1
+    for (int i = 0; i < controller.tabList.length - 1; i++) {
+      final String tag = controller.tabList[i];
+      _list.add(
+        Column(
+          key: ValueKey(tag),
+          children: [
+            _buildCupertinoListTile(controller.tabList[i]),
+            shortDivider,
+          ],
+        ),
+      );
+    }
+
+    _list.add(
+      _buildCupertinoListTile(controller.tabList.last),
+    );
+
+    return _list;
   }
 }
