--- lib/pages/setting_c/image_block/phash_list_page.dart
@@ -1,98 +1,0 @@
-import 'package:fehviewer/common/controller/image_hide_controller.dart';
-import 'package:fehviewer/fehviewer.dart';
-import 'package:fehviewer/widget/cupertino/sliver_list_section.dart';
-import 'package:flutter/cupertino.dart';
-import 'package:get/get.dart';
-
-class PHashImageListPage extends GetView<ImageHideController> {
-  const PHashImageListPage({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return CupertinoPageScaffold(
-      backgroundColor: CupertinoColors.systemGroupedBackground,
-      navigationBar: CupertinoNavigationBar(
-        padding: const EdgeInsetsDirectional.only(end: 12),
-        middle: Text(L10n.of(context).phash_block_list),
-        trailing: CupertinoButton(
-          // 清除按钮
-          padding: const EdgeInsets.all(0),
-          minSize: 40,
-          child: const Icon(
-            CupertinoIcons.trash,
-            size: 24,
-          ),
-          onPressed: () {
-            controller.customHideList.clear();
-          },
-        ),
-      ),
-      child: CustomScrollView(slivers: [
-        SliverSafeArea(
-          sliver: Obx(() {
-            return SliverCupertinoListSection.insetGrouped(
-              additionalDividerMargin: 64,
-              itemCount: controller.customHideList.length,
-              itemBuilder: (context, index) {
-                final imageHide = controller.customHideList[index];
-
-                return ImageHideItem(
-                  imageHide: imageHide,
-                  onDelete: () => controller.customHideList.removeAt(index),
-                );
-              },
-            );
-          }),
-        ),
-      ]),
-    );
-  }
-}
-
-class ImageHideItem extends StatelessWidget {
-  const ImageHideItem({Key? key, required this.imageHide, this.onDelete})
-      : super(key: key);
-  final ImageHide imageHide;
-  final VoidCallback? onDelete;
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      padding: const EdgeInsets.only(left: 14, right: 12),
-      margin: const EdgeInsets.symmetric(vertical: 8),
-      height: 80,
-      child: Row(
-        children: [
-          Expanded(
-            child: Row(
-              children: [
-                Padding(
-                  padding: const EdgeInsets.only(right: 12),
-                  child: ClipRRect(
-                    child: Container(
-                      width: 50,
-                      child: EhNetworkImage(imageUrl: imageHide.imageUrl ?? ''),
-                    ),
-                    borderRadius: BorderRadius.circular(8.0),
-                  ),
-                ),
-                Text(
-                  imageHide.pHash.toUpperCase(),
-                ),
-              ],
-            ),
-          ),
-          CupertinoButton(
-            // 清除按钮
-            child: const Icon(
-              CupertinoIcons.xmark_circle,
-              color: CupertinoColors.systemRed,
-              size: 26,
-            ),
-            onPressed: onDelete,
-          ),
-        ],
-      ),
-    );
-  }
-}
