+++ lib/pages/tab/view/home_page_large.dart
@@ -9,33 +9,33 @@ import 'package:fehviewer/pages/gallery/view/all_thumbnails_page.dart';
 import 'package:fehviewer/pages/gallery/view/comment_page.dart';
 import 'package:fehviewer/pages/gallery/view/gallery_info_page.dart';
 import 'package:fehviewer/pages/gallery/view/sliver/gallery_page.dart';
+import 'package:fehviewer/pages/setting/about_page.dart';
+import 'package:fehviewer/pages/setting/advanced_setting_page.dart';
 import 'package:fehviewer/pages/setting/avatar_setting_page.dart';
+import 'package:fehviewer/pages/setting/block/block_rule_edit_page.dart';
+import 'package:fehviewer/pages/setting/block/block_rules_page.dart';
+import 'package:fehviewer/pages/setting/block/blockers_page.dart';
 import 'package:fehviewer/pages/setting/controller/tab_setting_controller.dart';
 import 'package:fehviewer/pages/setting/custom_hosts_page.dart';
+import 'package:fehviewer/pages/setting/download_setting_page.dart';
 import 'package:fehviewer/pages/setting/eh_mysettings_page.dart';
+import 'package:fehviewer/pages/setting/eh_setting_page.dart';
+import 'package:fehviewer/pages/setting/image_block/phash_list_page.dart';
+import 'package:fehviewer/pages/setting/image_block_page.dart';
 import 'package:fehviewer/pages/setting/item_width_setting_page.dart';
+import 'package:fehviewer/pages/setting/layout_setting_page.dart';
+import 'package:fehviewer/pages/setting/license_page.dart';
+import 'package:fehviewer/pages/setting/log_page.dart';
 import 'package:fehviewer/pages/setting/mytags/eh_mytags_page.dart';
 import 'package:fehviewer/pages/setting/mytags/eh_usertag_page.dart';
 import 'package:fehviewer/pages/setting/proxy_page.dart';
+import 'package:fehviewer/pages/setting/read_setting_page.dart';
+import 'package:fehviewer/pages/setting/search_setting_page.dart';
+import 'package:fehviewer/pages/setting/security_setting_page.dart';
+import 'package:fehviewer/pages/setting/tabbar_setting_page.dart';
 import 'package:fehviewer/pages/setting/tag_translat_page.dart';
 import 'package:fehviewer/pages/setting/view/login_webdav.dart';
 import 'package:fehviewer/pages/setting/webdav_setting_page.dart';
-import 'package:fehviewer/pages/setting_c/about_page.dart';
-import 'package:fehviewer/pages/setting_c/advanced_setting_page.dart';
-import 'package:fehviewer/pages/setting_c/block/block_rule_edit_page.dart';
-import 'package:fehviewer/pages/setting_c/block/block_rules_page.dart';
-import 'package:fehviewer/pages/setting_c/block/blockers_page.dart';
-import 'package:fehviewer/pages/setting_c/download_setting_page.dart';
-import 'package:fehviewer/pages/setting_c/eh_setting_page.dart';
-import 'package:fehviewer/pages/setting_c/image_block/phash_list_page.dart';
-import 'package:fehviewer/pages/setting_c/image_block_page.dart';
-import 'package:fehviewer/pages/setting_c/layout_setting_page.dart';
-import 'package:fehviewer/pages/setting_c/license_page.dart';
-import 'package:fehviewer/pages/setting_c/log_page.dart';
-import 'package:fehviewer/pages/setting_c/read_setting_page.dart';
-import 'package:fehviewer/pages/setting_c/search_setting_page.dart';
-import 'package:fehviewer/pages/setting_c/security_setting_page.dart';
-import 'package:fehviewer/pages/setting_c/tabbar_setting_page.dart';
 import 'package:fehviewer/pages/tab/controller/tabhome_controller.dart';
 import 'package:fehviewer/pages/tab/view/download_page.dart';
 import 'package:fehviewer/pages/tab/view/quick_search_page.dart';
